Add files via upload
                                                                              HOTEL BOOKING DEMAND DATASET ANALYSIS.
INTRODUCTION: ABOUT THIS DATASET.
Have you ever wondered when the best time of the year to book a hotel room is? Or the optimal length of stay in order to get the best daily rate? What if you wanted to predict whether or not a hotel was likely to receive a disproportionately high number of special requests? This hotel booking dataset can help you explore those questions!

CONTENT:
The dataset for this project contains booking information for a City hotel and a Resort hotel, and includes information as when the booking was made, number of adults, children and/or babies, length of stay, and the number of available parking spaces, among other things.

DETAIL DESCRIPTION:
I have two criteria hotel demand data. One of the hotels is a Resort hotel (H1), while the other is a City hotel (H2). Both datasets share the same structure, with 31 variables describing the 40060 observations of H1 and 79330 observations of H2. Each observation represents a hotel booking.

In tourism and travel related industries, most of the research on Revenue Management demand forecasting and prediction problems.

However, the remaining tourism and travel industries like hospitality, cruising, theme parks, etc., have different requirements and peculiarities that cannot be fully explored without industry’s specific data. Hence, two hotel datasets with demand data are shared to help in overcoming this limitation.

The datasets now made available were collected aiming at the development of prediction models to classify a hotel booking’s likelihood to be canceled. Nevertheless, due to the characteristics of the variables included in these datasets, their use goes beyond this cancellation prediction problem. One of the most important properties in data for prediction models is not to promote leakage of future information

A word of caution is due for those not so familiar with hotel operations. In hotel industry, it is quite common for customers to change their booking’s attributes, like the number of persons, staying duration, or room type preferences, either at the time of their check-in or during their stay. It is also common for hotels not to know the correct nationality of the customer until the moment of check-in. Therefore, even though the capture of data took considered a timespan prior to arrival date, it is understandable that the distribution of some variables differ between non cancelled and cancelled bookings.

END-USER OBJECTIVE FOR HOTEL BOOKING DEMAND IS:
By utilizing the predictive model, hotels will be able to identify the reason why customers cancel their bookings and their reasons for doing so, also, what time of the year these cancelations are experience. It would be grateful to identify root cause and better strategy for hotel management teams.

I am going to extract insight about most of the customer prefer choice between Resort or City hotel.

By using Exploratory Data Analysis(EDA), I shall be analysing and visualising the different prospective, which will give us different insights into ( 1. What is the percentage of bookings for each year?, 2. Which is the busiest months?, 3. How many bookings were cancelled?, etc).

Visualizing more insights by performing Exploratory Data Analysis, I shall be helping hotels to get prepared with adequate and timely arrangements on heavy and low tourist visit periods.

IN SHORT,I WILL TRY TO PROVIDE ANSWERS TO THE FOLLOWING QUESTIONS FROM THE DATA:
1. How many bookings were cancelled?
2. What is the percentage of booking for each year?
3. Which is the busiest months for hotels?
4. From which country does most guest come from?
5. How long do guest stay in the hotel?
6. What is the booking ratio between Resort hotel and City hotel?
7. Which was the most booked accommodation type(single, couple, family)?
When I am done with providing answers to this questions, i will build a predictive model to make predictions in the future whether a booking will be cancelled or not.

I WILL:
1. Perform Data Exploration
2. Data Cleaning
3. Feature Engineering to Make New Features
4. Perform Feature Selection to Select Only Relevant Features
5. Transform the Data(Categorical to Numerical)
6. Split the Data(Train Test Split)
7. Model the Data(Fit the Data)
8. Evaluate the Model

TOOLS AND LIBRARIES USED:
We will use Python 3 and some of its packages:
1. Pandas
2. Matplotlib
3. Seaborn
4. Sklearn